@@ -96,14 +96,14 @@ return [
             'calendar' => [
                 'label' => _("txt-calendar-admin"),
                 'order' => 60,
-                'route' => 'zfcadmin/calendar-manager',
+                'route' => 'zfcadmin/calendar',
                 'pages' => [
                     'calendar'          => [
                         'label' => _("txt-calendar"),
-                        'route' => 'zfcadmin/calendar-manager/overview',
+                        'route' => 'zfcadmin/calendar/overview',
                         'pages' => [
                             'view-calendar' => [
-                                'route'   => 'zfcadmin/calendar-manager/calendar',
+                                'route'   => 'zfcadmin/calendar/calendar',
                                 'visible' => false,
                                 'params'  => [
                                     'entities'   => [
@@ -116,7 +116,7 @@ return [
                                 'pages'   => [
                                     'edit-calendar'    => [
                                         'label'   => _("txt-edit-calendar"),
-                                        'route'   => 'zfcadmin/calendar-manager/edit',
+                                        'route'   => 'zfcadmin/calendar/edit',
                                         'visible' => false,
                                         'params'  => [
                                             'entities' => [
@@ -126,7 +126,7 @@ return [
                                     ],
                                     'select-attendees' => [
                                         'label'   => _("txt-nav-calendar-select-attendees"),
-                                        'route'   => 'zfcadmin/calendar-manager/select-attendees',
+                                        'route'   => 'zfcadmin/calendar/select-attendees',
                                         'visible' => false,
                                         'params'  => [
                                             'entities' => [
@@ -140,7 +140,7 @@ return [
                     ],
                     'new-calendar-item' => [
                         'label' => _("txt-add-calendar-item"),
-                        'route' => 'zfcadmin/calendar-manager/new',
+                        'route' => 'zfcadmin/calendar/new',
                     ],
                 ],
             ],
