@@ -18,7 +18,7 @@ use Zend\Permissions\Acl\Acl;
 use Zend\Permissions\Acl\Resource\ResourceInterface;
 use Zend\Permissions\Acl\Role\RoleInterface;
 
-class Contact extends AssertionAbstract
+class Contact extends AbstractAssertion
 {
     /**
      * Returns true if and only if the assertion conditions are met.
@@ -27,29 +27,37 @@ class Contact extends AssertionAbstract
      * $role, $contact, or $privilege parameters are null, it means that the query applies to all Roles, Resources, or
      * privileges, respectively.
      *
-     * @param Acl $acl
-     * @param RoleInterface $role
+     * @param Acl               $acl
+     * @param RoleInterface     $role
      * @param ResourceInterface $contact
-     * @param string $privilege
+     * @param string            $privilege
      *
      * @return bool
      */
-    public function assert(Acl $acl, RoleInterface $role = null, ResourceInterface $contact = null, $privilege = null)
-    {
+    public function assert(
+        Acl $acl,
+        RoleInterface $role = null,
+        ResourceInterface $contact = null,
+        $privilege = null
+    ): bool {
         $this->setPrivilege($privilege);
         $id = $this->getId();
 
-        if (!$contact instanceof ContactEntity) {
-            $contact = $this->getCalendarService()->findEntityById(ContactEntity::class, $id);
+        if (!$contact instanceof ContactEntity && null !== $id) {
+            $contact = $this->calendarService->find(ContactEntity::class, $id);
+        }
+
+        if (null === $contact) {
+            return true;
         }
 
         switch ($this->getPrivilege()) {
             case 'update-status':
-                if ($this->getCalendarService()->calendarHasContact($contact->getCalendar(), $this->getContact())) {
+                if ($this->calendarService->calendarHasContact($contact->getCalendar(), $this->contact)) {
                     return true;
                 }
 
-                return $this->rolesHaveAccess([Access::ACCESS_OFFICE]);
+                return $this->rolesHaveAccess(Access::ACCESS_OFFICE);
         }
 
         return false;
