@@ -18,7 +18,7 @@ use Zend\Permissions\Acl\Acl;
 use Zend\Permissions\Acl\Resource\ResourceInterface;
 use Zend\Permissions\Acl\Role\RoleInterface;
 
-class Document extends AssertionAbstract
+class Document extends AbstractAssertion
 {
     /**
      * Returns true if and only if the assertion conditions are met.
@@ -27,10 +27,10 @@ class Document extends AssertionAbstract
      * $role, $document, or $privilege parameters are null, it means that the query applies to all Roles, Resources, or
      * privileges, respectively.
      *
-     * @param Acl $acl
-     * @param RoleInterface $role
+     * @param Acl               $acl
+     * @param RoleInterface     $role
      * @param ResourceInterface $document
-     * @param string $privilege
+     * @param string            $privilege
      *
      * @return bool
      */
@@ -39,14 +39,13 @@ class Document extends AssertionAbstract
         RoleInterface $role = null,
         ResourceInterface $document = null,
         $privilege = null
-    ) {
+    ): bool {
         $this->setPrivilege($privilege);
         $id = $this->getId();
 
-        if (!$document instanceof DocumentEntity) {
-            $document = $this->getCalendarService()->findEntityById(DocumentEntity::class, $id);
+        if (!$document instanceof DocumentEntity && null !== $id) {
+            $document = $this->calendarService->find(DocumentEntity::class, $id);
         }
-
 
         /*
          * No document was found, so return true because we do not now anything about the access
@@ -58,28 +57,22 @@ class Document extends AssertionAbstract
         switch ($this->getPrivilege()) {
             case 'document-community':
             case 'download':
-                if ($this->getContactService()->contactHasPermit(
-                    $this->getContact(),
-                    'view',
-                    $document->getCalendar()
-                )) {
+                if ($this->hasPermission($document->getCalendar(), 'view')) {
                     return true;
                 }
 
-                return $this->getCalendarService()->canViewCalendar($document->getCalendar(), $this->getContact());
+                return $this->calendarService->canViewCalendar($document->getCalendar(), $this->contact);
             case 'edit-community':
-                if ($this->getContactService()
-                    ->contactHasPermit($this->getContact(), 'edit', $document->getCalendar())
-                ) {
+                if ($this->hasPermission($document->getCalendar(), 'edit')) {
                     return true;
                 }
 
                 /*
                  * The project leader also has rights to invite users
                  */
-                if (!\is_null($document->getCalendar()->getProjectCalendar())) {
-                    if ($this->getContactService()->contactHasPermit(
-                        $this->getContact(),
+                if (null !== $document->getCalendar()->getProjectCalendar()) {
+                    if ($this->contactService->contactHasPermit(
+                        $this->contact,
                         'edit',
                         $document->getCalendar()->getProjectCalendar()->getProject()
                     )
@@ -88,9 +81,9 @@ class Document extends AssertionAbstract
                     }
                 }
 
-                return $this->rolesHaveAccess([Access::ACCESS_OFFICE]);
+                return $this->rolesHaveAccess(Access::ACCESS_OFFICE);
             case 'document-admin':
-                return $this->rolesHaveAccess([Access::ACCESS_OFFICE]);
+                return $this->rolesHaveAccess(Access::ACCESS_OFFICE);
         }
 
         return false;
