@@ -15,51 +15,104 @@ namespace Calendar\Controller;
 use Calendar\Entity\Document;
 use Calendar\Entity\DocumentObject;
 use Calendar\Form\CreateCalendarDocument;
+use Calendar\Service\CalendarService;
+use Doctrine\ORM\EntityManager;
+use General\Service\GeneralService;
+use Zend\Http\Response;
+use Zend\I18n\Translator\TranslatorInterface;
+use Zend\Mvc\Controller\AbstractActionController;
+use Zend\Mvc\Plugin\FlashMessenger\FlashMessenger;
 use Zend\Validator\File\FilesSize;
 use Zend\Validator\File\MimeType;
 use Zend\View\Model\ViewModel;
 
 /**
+ * Class DocumentController
  *
+ * @package Calendar\Controller
+ * @method FlashMessenger flashMessenger()
  */
-class CalendarDocumentController extends CalendarAbstractController
+class DocumentController extends AbstractActionController
 {
     /**
-     * @return \Zend\Stdlib\ResponseInterface|ViewModel
+     * @var CalendarService
      */
-    public function downloadAction()
+    protected $calendarService;
+    /**
+     * @var GeneralService
+     */
+    protected $generalService;
+    /**
+     * @var EntityManager
+     */
+    protected $entityManager;
+    /**
+     * @var TranslatorInterface
+     */
+    protected $translator;
+
+    /**
+     * DocumentController constructor.
+     *
+     * @param CalendarService     $calendarService
+     * @param GeneralService      $generalService
+     * @param EntityManager       $entityManager
+     * @param TranslatorInterface $translator
+     */
+    public function __construct(
+        CalendarService $calendarService,
+        GeneralService $generalService,
+        EntityManager $entityManager,
+        TranslatorInterface $translator
+    ) {
+        $this->calendarService = $calendarService;
+        $this->generalService = $generalService;
+        $this->entityManager = $entityManager;
+        $this->translator = $translator;
+    }
+
+
+    /**
+     * @return Response
+     */
+    public function downloadAction(): Response
     {
+        /** @var Response $response */
+        $response = $this->getResponse();
+
         /**
          * @var Document $document
          */
-        $document = $this->getCalendarService()->findEntityById(Document::class, $this->params('id'));
-        if (\is_null($document) || count($document->getObject()) === 0) {
-            return $this->notFoundAction();
+        $document = $this->calendarService->find(Document::class, (int)$this->params('id'));
+
+        if (null === $document || \count($document->getObject()) === 0) {
+            return $response->setStatusCode(Response::STATUS_CODE_404);
         }
 
         /*
          * Due to the BLOB issue, we treat this as an array and we need to capture the first element
          */
         $object = $document->getObject()->first()->getObject();
-        $response = $this->getResponse();
+
         $response->setContent(stream_get_contents($object));
-        $response->getHeaders()->addHeaderLine('Expires: ' . gmdate('D, d M Y H:i:s \G\M\T', time() + 36000))
-            ->addHeaderLine("Cache-Control: max-age=36000, must-revalidate")
+        $response->getHeaders()
+            ->addHeaderLine('Expires: ' . gmdate('D, d M Y H:i:s \G\M\T', time() + 36000))
+            ->addHeaderLine('Cache-Control: max-age=36000, must-revalidate')
             ->addHeaderLine('Content-Disposition', 'attachment; filename="' . $document->parseFileName() . '"')
-            ->addHeaderLine("Pragma: public")->addHeaderLine(
+            ->addHeaderLine('Pragma: public')->addHeaderLine(
                 'Content-Type: ' . $document->getContentType()
                     ->getContentType()
             )->addHeaderLine('Content-Length: ' . $document->getSize());
 
-        return $this->response;
+        return $response;
     }
 
     /**
-     * @return array|ViewModel
+     * @return ViewModel
      */
-    public function documentAction()
+    public function documentAction(): ViewModel
     {
-        $document = $this->getCalendarService()->findEntityById(Document::class, $this->params('id'));
+        $document = $this->calendarService->find(Document::class, (int)$this->params('id'));
 
         if (null === $document) {
             return $this->notFoundAction();
@@ -69,12 +122,14 @@ class CalendarDocumentController extends CalendarAbstractController
     }
 
     /**
-     * @return array|\Zend\Http\Response|ViewModel
+     * @return Response|ViewModel
+     * @throws \Doctrine\ORM\ORMException
+     * @throws \Doctrine\ORM\OptimisticLockException
      */
     public function editAction()
     {
         /** @var Document $document */
-        $document = $this->getCalendarService()->findEntityById(Document::class, $this->params('id'));
+        $document = $this->calendarService->find(Document::class, (int)$this->params('id'));
         if (null === $document) {
             return $this->notFoundAction();
         }
@@ -83,7 +138,7 @@ class CalendarDocumentController extends CalendarAbstractController
             $this->getRequest()->getPost()->toArray(),
             $this->getRequest()->getFiles()->toArray()
         );
-        $form = new CreateCalendarDocument($this->getEntityManager());
+        $form = new CreateCalendarDocument($this->entityManager);
         $form->bind($document);
         $form->getInputFilter()->get('file')->setRequired(false);
         $form->setData($data);
@@ -99,11 +154,11 @@ class CalendarDocumentController extends CalendarAbstractController
                 $this->flashMessenger()->setNamespace('success')
                     ->addMessage(
                         sprintf(
-                            $this->translate("txt-calendar-document-%s-successfully-removed"),
+                            $this->translator->translate("txt-calendar-document-%s-successfully-removed"),
                             $document->parseFileName()
                         )
                     );
-                $this->getCalendarService()->removeEntity($document);
+                $this->calendarService->delete($document);
 
                 return $this->redirect()
                     ->toRoute(
@@ -133,7 +188,9 @@ class CalendarDocumentController extends CalendarAbstractController
 
                     $fileTypeValidator = new MimeType();
                     $fileTypeValidator->isValid($file);
-                    $document->setContentType($this->getGeneralService()->findContentTypeByContentTypeName($fileTypeValidator->type));
+                    $document->setContentType(
+                        $this->generalService->findContentTypeByContentTypeName($fileTypeValidator->type)
+                    );
 
                     /**
                      * Update the object
@@ -142,13 +199,13 @@ class CalendarDocumentController extends CalendarAbstractController
                      */
                     $documentObject = $document->getObject()->first();
                     $documentObject->setObject(file_get_contents($file['tmp_name']));
-                    $this->getCalendarService()->updateEntity($documentObject);
+                    $this->calendarService->save($documentObject);
                 }
-                $this->getCalendarService()->updateEntity($document);
+                $this->calendarService->save($document);
                 $this->flashMessenger()->setNamespace('success')
                     ->addMessage(
                         sprintf(
-                            $this->translate("txt-calendar-document-%s-successfully-updated"),
+                            $this->translator->translate("txt-calendar-document-%s-successfully-updated"),
                             $document->parseFileName()
                         )
                     );
