@@ -1,27 +1,27 @@
 <?php
 // Generated by ZF2's ./bin/templatemap_generator.php
 return array(
-    'calendar/pdf/review-calendar'                 => __DIR__ . '//view/calendar/pdf/review-calendar.twig',
-    'calendar/pdf/signature-list'                  => __DIR__ . '//view/calendar/pdf/signature-list.twig',
-    'calendar/pdf/presence-list'                   => __DIR__ . '//view/calendar/pdf/presence-list.twig',
-    'calendar/calendar/calendar-type-color-css'    => __DIR__ . '//view/calendar/calendar/calendar-type-color-css.twig',
-    'calendar/partial/pagination-control'          => __DIR__ . '//view/calendar/partial/pagination-control.twig',
-    'calendar/partial/entity/calendar'             => __DIR__ . '//view/calendar/partial/entity/calendar.twig',
-    'calendar/partial/list/calendar-contact'       => __DIR__ . '//view/calendar/partial/list/calendar-contact.twig',
-    'calendar/partial/list/calendar-community'     => __DIR__ . '//view/calendar/partial/list/calendar-community.twig',
-    'calendar/partial/year-selector'               => __DIR__ . '//view/calendar/partial/year-selector.twig',
-    'calendar/partial/form/calendar'               => __DIR__ . '//view/calendar/partial/form/calendar.twig',
-    'calendar/calendar-document/edit'              => __DIR__ . '//view/calendar/calendar-document/edit.twig',
-    'calendar/calendar-document/document'          => __DIR__ . '//view/calendar/calendar-document/document.twig',
-    'calendar/calendar-community/review-calendar'  => __DIR__ . '//view/calendar/calendar-community/review-calendar.twig',
-    'calendar/calendar-community/calendar'         => __DIR__ . '//view/calendar/calendar-community/calendar.twig',
-    'calendar/calendar-community/overview'         => __DIR__ . '//view/calendar/calendar-community/overview.twig',
-    'calendar/calendar-community/contact'          => __DIR__ . '//view/calendar/calendar-community/contact.twig',
-    'calendar/calendar-community/send-message'     => __DIR__ . '//view/calendar/calendar-community/send-message.twig',
-    'calendar/calendar-community/select-attendees' => __DIR__ . '//view/calendar/calendar-community/select-attendees.twig',
-    'calendar/calendar-manager/edit'               => __DIR__ . '//view/calendar/calendar-manager/edit.twig',
-    'calendar/calendar-manager/calendar'           => __DIR__ . '//view/calendar/calendar-manager/calendar.twig',
-    'calendar/calendar-manager/new'                => __DIR__ . '//view/calendar/calendar-manager/new.twig',
-    'calendar/calendar-manager/overview'           => __DIR__ . '//view/calendar/calendar-manager/overview.twig',
-    'calendar/calendar-manager/select-attendees'   => __DIR__ . '//view/calendar/calendar-manager/select-attendees.twig',
+    'calendar/pdf/review-calendar'              => __DIR__ . '//view/calendar/pdf/review-calendar.twig',
+    'calendar/pdf/signature-list'               => __DIR__ . '//view/calendar/pdf/signature-list.twig',
+    'calendar/pdf/presence-list'                => __DIR__ . '//view/calendar/pdf/presence-list.twig',
+    'calendar/calendar/calendar-type-color-css' => __DIR__ . '//view/calendar/calendar/calendar-type-color-css.twig',
+    'calendar/partial/pagination-control'       => __DIR__ . '//view/calendar/partial/pagination-control.twig',
+    'calendar/partial/entity/calendar'          => __DIR__ . '//view/calendar/partial/entity/calendar.twig',
+    'calendar/partial/list/calendar-contact'    => __DIR__ . '//view/calendar/partial/list/calendar-contact.twig',
+    'calendar/partial/list/calendar-community'  => __DIR__ . '//view/calendar/partial/list/calendar-community.twig',
+    'calendar/partial/year-selector'            => __DIR__ . '//view/calendar/partial/year-selector.twig',
+    'calendar/partial/form/calendar'            => __DIR__ . '//view/calendar/partial/form/calendar.twig',
+    'calendar/document/edit'                    => __DIR__ . '//view/calendar/document/edit.twig',
+    'calendar/document/document'                => __DIR__ . '//view/calendar/document/document.twig',
+    'calendar/community/review-calendar'        => __DIR__ . '//view/calendar/community/review-calendar.twig',
+    'calendar/community/calendar'               => __DIR__ . '//view/calendar/community/calendar.twig',
+    'calendar/community/overview'               => __DIR__ . '//view/calendar/community/overview.twig',
+    'calendar/community/contact'                => __DIR__ . '//view/calendar/community/contact.twig',
+    'calendar/community/send-message'           => __DIR__ . '//view/calendar/community/send-message.twig',
+    'calendar/community/select-attendees'       => __DIR__ . '//view/calendar/community/select-attendees.twig',
+    'calendar/manager/edit'                     => __DIR__ . '//view/calendar/manager/edit.twig',
+    'calendar/manager/calendar'                 => __DIR__ . '//view/calendar/manager/calendar.twig',
+    'calendar/manager/new'                      => __DIR__ . '//view/calendar/manager/new.twig',
+    'calendar/manager/overview'                 => __DIR__ . '//view/calendar/manager/overview.twig',
+    'calendar/manager/select-attendees'         => __DIR__ . '//view/calendar/manager/select-attendees.twig',
 );
